# specs/001-mvp-closed-loop/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PPTagent MVP API
  version: 0.1.0-mvp
  description: |
    REST API for PPTagent MVP 业务闭环（生成 / 知识库 / Agent 进化）。
    与 WebSocket 事件契约（events.yaml）配合使用，错误码遵循 RFC 7807。
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev
  - url: https://api.pptagent.example.com/api/v1
    description: Production (placeholder)
security:
  - ApiKeyAuth: []
paths:
...
  /drafts:
    get:
      summary: 列出当前用户的活跃草稿
      operationId: listDrafts
      tags: [Drafts]
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [active, archived, exported] }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftPage' }
    post:
      summary: 创建草稿
      operationId: createDraft
      tags: [Drafts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDraftRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftDetail' }
  /drafts/{draft_id}:
    parameters:
      - in: path
        name: draft_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 草稿详情（含 slides 列表）
      operationId: getDraft
      tags: [Drafts]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: 更新草稿（标题 / 整体风格 / 自动保存；FR-035）
      operationId: updateDraft
      tags: [Drafts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDraftRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftDetail' }
        '409':
          description: 乐观锁失败（expected_revision 与服务端不一致）
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '423':
          description: 草稿被其他用户锁定
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    delete:
      summary: 归档草稿（soft delete → status=archived）
      operationId: archiveDraft
      tags: [Drafts]
      responses:
        '204': { description: OK }
  /drafts/{draft_id}/lock:
    parameters:
      - in: path
        name: draft_id
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 获取草稿编辑锁（FR-035 单写者；30 分钟自动过期）
      operationId: acquireDraftLock
      tags: [Drafts]
      responses:
        '200':
          description: 取得锁
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftLockInfo' }
        '423':
          description: 锁被他人持有
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    delete:
      summary: 主动释放草稿编辑锁
      operationId: releaseDraftLock
      tags: [Drafts]
      responses:
        '204': { description: OK }
  /drafts/{draft_id}/slides/{slide_id}:
    parameters:
      - in: path
        name: draft_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: slide_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: 调整草稿单张页（顺序 / 样式覆盖 / 文本）
      operationId: updateDraftSlide
      tags: [Drafts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDraftSlideRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftSlide' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: 删除草稿中的某张页
      operationId: deleteDraftSlide
      tags: [Drafts]
      responses:
        '204': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
  /drafts/{draft_id}/export:
    parameters:
      - in: path
        name: draft_id
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: 导出草稿为 PPTX（FR-036 保留来源标注为 XMP / 自定义 XML）
      operationId: exportDraft
      tags: [Drafts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DraftExportRequest' }
      responses:
        '202':
          description: 已入队打包
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftExportJob' }
        '423': { $ref: '#/components/responses/Locked' }
  /drafts/{draft_id}/export/{job_id}:
    parameters:
      - in: path
        name: draft_id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: job_id
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 查询草稿导出任务进度
      operationId: getDraftExportJob
      tags: [Drafts]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftExportJob' }

  # ─── Preferences ───
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 形如 `Bearer <key>`（最小权限 IV + 速率限制）
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 7807
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string, description: PPTagent 自定义错误码，详见 error-codes.yaml }
        request_id: { type: string, format: uuid, description: 关联 X-Request-Id }
        trace_id: { type: string, description: OTel trace ID }
        errors:
          type: array
          description: 字段级错误（422 验证失败时填充）
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }

    # ─── Pagination ───
...
    DraftStatus:
      type: string
      enum: [active, archived, exported]
    DraftSlideSourceType:
      type: string
      enum: [reused, generated, manual]
...
    Draft:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { $ref: '#/components/schemas/DraftStatus' }
        overall_style: { type: object, additionalProperties: true, nullable: true }
        editor_user_id: { type: string, format: uuid, nullable: true }
        last_saved_at: { type: string, format: date-time }
        slide_count: { type: integer }
        created_at: { type: string, format: date-time }
        revision: { type: integer, description: 乐观锁当前值 }
    DraftSlide:
      type: object
      properties:
        id: { type: string, format: uuid }
        slide_order: { type: integer, minimum: 1 }
        source_type: { $ref: '#/components/schemas/DraftSlideSourceType' }
        source_asset_id: { type: string, format: uuid, nullable: true }
        source_stage_id: { type: string, format: uuid, nullable: true }
        thumbnail_url: { type: string, nullable: true }
        style_normalized: { type: boolean }
        normalized_failed: { type: boolean }
        source_attribution: { type: string }
    DraftDetail:
      allOf:
        - $ref: '#/components/schemas/Draft'
        - type: object
          properties:
            slides:
              type: array
              items: { $ref: '#/components/schemas/DraftSlide' }
    DraftPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Draft' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        has_next: { type: boolean }
    CreateDraftRequest:
      type: object
      properties:
        title: { type: string, maxLength: 256, default: 'Untitled Draft' }
        initial_style:
          type: object
          additionalProperties: true
          nullable: true
          description: 可选：手动指定整体风格（color_palette / font_family / layout_id）
    UpdateDraftRequest:
      type: object
      required: [expected_revision]
      properties:
        title: { type: string, maxLength: 256 }
        overall_style: { type: object, additionalProperties: true, nullable: true }
        slides_snapshot:
          type: array
          items: { $ref: '#/components/schemas/DraftSlide' }
          description: 客户端完整快照；服务端 diff 后写回
        expected_revision: { type: integer, minimum: 0 }
    UpdateDraftSlideRequest:
      type: object
      properties:
        slide_order: { type: integer, minimum: 1 }
        style_overrides: { type: object, additionalProperties: true, nullable: true }
        text_overrides: { type: object, additionalProperties: true, nullable: true }
        expected_revision: { type: integer, minimum: 0 }
    DraftLockInfo:
      type: object
      properties:
        editor_user_id: { type: string, format: uuid }
        acquired_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time, description: 30 分钟后自动过期 }
    DraftExportRequest:
      type: object
      required: [expected_revision]
      properties:
        include_source_attribution: { type: boolean, default: true, description: FR-036 写入 XMP / 自定义 XML }
        final_style_overrides: { type: object, additionalProperties: true, nullable: true }
        expected_revision: { type: integer, minimum: 0 }
    DraftExportJob:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, running, ready, failed] }
        download_url: { type: string, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true }
        error:
          type: object
          additionalProperties: true
          nullable: true